Add support for using strings when declaring the model for an embedment
DM::M::Embedments::Relationship will now accept a string instead of
an EmbeddedModel, and will attempt to find the constant the first
time Relationship#target_model is called.

Lazily fetching the constant for the source_model isn't needed since
we always know the source model when the relationship is created (as
there is no support for ManyToOne relationships).